Fix some problems with API docs.
As I noted on the mailing list, I found some inconsistencies which
seemed to be cut-and=paste errors, and as I started to describe them, I
realized it would be nicer to just submit a pull request. Since I was
doing that, I expanded my efforts to include making the spacing
consistent between them, and adding some additional disambiguation that
I thought would be helpful.